requirements.txt
toml==0.10.2
librosa==0.10.2.post1
open-clip-torch==2.24.0
pytorch-msssim==1.0.0